+++ NEWS.md
@@ -1,6 +1,9 @@
 # lavaanExtra 0.1.x
 * Work in progress!
 
+## lavaanExtra 0.1.4.6
+* `nice_tidySEM`: now omit the leading zero when using standardized coefficients (for APA style).
+
 ## lavaanExtra 0.1.4.5
 * `lavaan_cov`: renamed r column to covariance since in some cases standardized residual variances were not real correlations.
 * New function: `lavaan_cor`, which is the same as `lavaan_cov` but only for actual correlations